+++ app/src/Calculator/components/CalculatorInput/CalculatorInputMobile.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { useSetAtom } from 'jotai';
 import { Fragment } from 'react';
 
-import { OrderItemButtonGroup } from '@/Calculator/components/OrderItemButton';
+import { OrderItemButtonGroup } from '@/Calculator/components/OrderItemButtonGroup';
 import { ProjectSpotlight } from '@/Calculator/components/ProjectSpotlight';
 import { useSubjectList } from '@/Calculator/hooks/useSubjectList';
 import type {
